move recordattributes table to its own file
